refactor: logging 방식을 parameterized 형식으로 변경
