+++ src/main/resources/application.properties
@@ -1,7 +1,8 @@
 spring.application.name=newsfeedApp
 
 spring.datasource.url=jdbc:mysql://localhost:3306/newsfeed
-spring.datasource.username=root
+spring.datasource.username=${LOCAL_DB_NAME}
+spring.datasource.password=${LOCAL_DB_PASSWORD}
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
 spring.jpa.hibernate.ddl-auto=update
